Relocate `PublishableFluentContentsAdmin` into publishing app, re #22
- move `PublishableFluentContentsAdmin` class from `articles` models to
  the central publishing app location that makes more sense
- relocate `FluentLayoutsMixin` admin mixin from `icekit.admin` module
  to the new `admin_mixins` module to avoid circular import problems
- fix imports and document references to models and mixin to refer to
  new location.
