+++ icekit/admin.py
@@ -10,9 +10,6 @@ from django.contrib import admin
 from django.contrib.contenttypes.models import ContentType
 from django.http import JsonResponse
 from django.utils.translation import ugettext_lazy as _
-from fluent_contents.admin import PlaceholderEditorAdmin
-from fluent_contents.analyzer import get_template_placeholder_data
-from fluent_contents.models import PlaceholderData
 from polymorphic.admin import PolymorphicParentModelAdmin
 
 from icekit import models
@@ -42,37 +39,6 @@ class ChildModelFilter(admin.SimpleListFilter):
 
 
 # MIXINS ######################################################################
-
-
-class FluentLayoutsMixin(PlaceholderEditorAdmin):
-    """
-    Mixin class for models that have a ``layout`` field and fluent content.
-    """
-
-    change_form_template = 'icekit/admin/fluent_layouts_change_form.html'
-
-    class Media:
-        js = ('icekit/admin/js/fluent_layouts.js', )
-
-    def formfield_for_foreignkey(self, db_field, *args, **kwargs):
-        """
-        Update queryset for ``layout`` field.
-        """
-        formfield = super(FluentLayoutsMixin, self).formfield_for_foreignkey(
-            db_field, *args, **kwargs)
-        if db_field.name == 'layout':
-            formfield.queryset = formfield.queryset.for_model(self.model)
-        return formfield
-
-    def get_placeholder_data(self, request, obj):
-        """
-        Get placeholder data from layout.
-        """
-        if not obj or not obj.layout:
-            data = [PlaceholderData(slot='main', role='m', title='Main')]
-        else:
-            data = obj.layout.get_placeholder_data()
-        return data
 
 
 class ChildModelPluginPolymorphicParentModelAdmin(PolymorphicParentModelAdmin):
