+++ icekit/articles/admin.py
@@ -1,14 +1,6 @@
 from django.contrib import admin
 
-from icekit.admin import FluentLayoutsMixin
-from icekit.publishing.admin import PublishingAdmin
-
-
-class PublishableFluentContentsAdmin(PublishingAdmin, FluentLayoutsMixin):
-    """
-    Add publishing features for non-Page rich content models
-    """
-    pass
+from icekit.publishing.admin import PublishableFluentContentsAdmin
 
 
 class TitleSlugAdmin(admin.ModelAdmin):
