+++ icekit/page_types/layout_page/admin.py
@@ -1,6 +1,6 @@
 from fluent_pages.integration.fluent_contents.admin import FluentContentsPageAdmin
 
-from icekit.admin import FluentLayoutsMixin
+from icekit.admin_mixins import FluentLayoutsMixin
 from icekit.publishing.admin import PublishingAdmin
